Выполняет шаги из задания - Личный проект: капитан ООП
